fix line preview style
