+++ GV_Core_with_OpenGL/factor.hpp
@@ -27,4 +27,5 @@ void viewScroll(GLFWwindow* window, double xoffset, double yoffset);
 void processCursorTrace(GLFWwindow* window,double xpos, double ypos);
 bool primitiveSelectDetect(Primitive* primitive);
 void generateNewPrimitive();
+void generatePreviewPrimitive(const vertexArray&);
 #endif /* factor_hpp */
